break job env handling into 3 parts:   - jobenv: moab, slurm, cobalt, openmpi, mpich, etc.   - jobruncli: aprun, srun, openmpi, mpich, etc.  [command line interface,          i.e. the set of flags provided]   - jobruncmd: actual name of jobruncmd... often the same as jobruncli          but can be different (e.g. "mpirun" vs "mpirun.mpich")
rework common init to to fit the above.  allow user to override with
environment variables (mainly for testing).  also for systems layered
directly on mpi, allow user to specify the mpirun command with $MPIRUN.

cobalt update: some cobalts use $COBALT_PARTNAME (theta) and others
use $COBALT_NODEFILE (cooley) to provide the list of compute nodes.
we hide the differences in interfaces with the cobalt_pbs_nodefile
script (adjust code here for that).

when looking for the environment, the priority order for doing so is:
1. allow users to hardwire the jobenv using JOBENV
2. look for job ID#s from a scheduler (moab, slurm, cobalt) to do setup
3. look to see if we are already running under openmpi or mpich
    otherwise we look for a bare MPI environment like this:
4. allow users to use $MPIRUN to specify a specific MPI to use
5. look for mpirun.mpich
6. look for mpirun.openmpi
7. look for plain mpirun and see if it is mpich or openmpi
8. give up

XXX: we need to move all the vpic-specific stuff out into its
own file so that common.sh is application independent